requirements.txt
discord.py==2.3.2
fastapi==0.109.0
gradio==4.15.0
python-dotenv==1.0.0
requests==2.31.0
Telethon==1.33.1
uvicorn==0.26.0
aiofiles==23.2.1
loguru==0.7.2